requirements.txt
pynput>=1.8.0
pystray>=0.19.0
cryptography>=42.0.0
argon2-cffi>=23.1.0
pyperclip>=1.9.0
Pillow>=10.0.0
